refactor: organize files to be more familiar to users coming from next.js
