+++ README.md
@@ -33,7 +33,7 @@ The browser version is in the `dist/client` folder and the Node.js version is in
 
 1. [package.json](package.json) - Start by reading the `scripts` section to understand how the build process works. Yes it's that dead simple!
 1. [app/index.html](app/index.html) - Your client-side HTML file (for development)
-1. [app/index.html](app/index.html) - Your server-side HTML file template (for server rendering)
+1. [server/template.js](server/template.js) - Your server-side HTML file template (for server rendering)
 1. [app/client.js](app/client.js) - Entry point for your browser version
 1. [app/App.jsx](app/App.jsx) - Your main application component shared between your browser and Node.js version
 1. [app/HelloWorld.jsx](app/HelloWorld.jsx) - Dead simple hello world
